move Devices to Mining namespace
