+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Devices.cs
@@ -1,4 +1,4 @@
-using NiceHashMiner.Devices;
+using NiceHashMiner.Mining;
 using NiceHashMiner.Stats;
 using System.Collections.Generic;
 using System.Linq;
