+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Mining.cs
@@ -1,4 +1,3 @@
-using NiceHashMiner.Devices;
 using System.Collections.Generic;
 using System.Linq;
 using NHM.Common.Enums;
