+++ src/NiceHashMiner/Mining/Grouping/GroupSetupUtils.cs
@@ -1,5 +1,4 @@
-﻿using NiceHashMiner.Devices;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
